refactor: Abstract feedStatus and order jobs and gigs
